Add sample 05
