+++ crates/piet-wgpu/examples/basics_wgpu.rs
@@ -17,28 +17,18 @@
 
 //! The `piet-glow` basics.rs example, but with piet-wgpu.
 
-use futures_lite::future;
-use instant::{Duration, Instant};
+#[path = "util/setup_context.rs"]
+mod util;
 
-use winit::event::{Event, WindowEvent};
-use winit::event_loop::EventLoop;
-use winit::window::WindowBuilder;
+use instant::{Duration, Instant};
 
 use piet_hardware::piet::kurbo::{Affine, BezPath, Point, Rect, Vec2};
 use piet_hardware::piet::{self, GradientStop, RenderContext as _};
-use piet_wgpu::{RenderContext, WgpuContext};
+use piet_wgpu::RenderContext;
 
 const ORANGES: &[u8] = include_bytes!("../../piet-glow/examples/assets/oranges.jpg");
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tracing_subscriber::fmt::init();
-    let event_loop = EventLoop::new();
-
-    let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
-        backends: wgpu::Backends::PRIMARY,
-        dx12_shader_compiler: Default::default(),
-    });
-
     // A path representing a star.
     let star = generate_five_pointed_star(Point::new(0.0, 0.0), 75.0, 150.0);
     let mut tick = 0;
@@ -60,7 +50,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut num_frames = 0;
     let mut current_fps = None;
 
-    let mut draw = move |rc: &mut RenderContext<'_, '_, '_>, width, height| {
+    let draw = move |rc: &mut RenderContext<'_, '_, '_>, width, height| {
         rc.clear(None, piet::Color::rgb8(0x87, 0xCE, 0xEB));
 
         let red_star = {
@@ -201,131 +191,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         rc.finish().unwrap();
     };
 
-    let mut state = None;
-    let format = wgpu::TextureFormat::Bgra8Unorm;
-    let mut config = wgpu::SurfaceConfiguration {
-        usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
-        format,
-        width: 0,
-        height: 0,
-        present_mode: wgpu::PresentMode::AutoVsync,
-        alpha_mode: wgpu::CompositeAlphaMode::Opaque,
-        view_formats: vec![format],
-    };
-    let mut window_size = (0, 0);
-
-    event_loop.run(move |ev, elwt, control_flow| {
-        control_flow.set_poll();
-
-        match ev {
-            Event::Resumed => {
-                let window = WindowBuilder::new()
-                    .with_title("piet-wgpu basics")
-                    .build(elwt)
-                    .expect("Failed to create window");
-
-                let size = window.inner_size();
-
-                let surface =
-                    unsafe { instance.create_surface(&window) }.expect("Failed to create surface");
-
-                let adaptor =
-                    future::block_on(instance.request_adapter(&wgpu::RequestAdapterOptions {
-                        compatible_surface: Some(&surface),
-                        ..Default::default()
-                    }))
-                    .expect("Failed to find an appropriate adapter");
-
-                // Create the logical device and command queue
-                let (device, queue) = future::block_on(adaptor.request_device(
-                    &wgpu::DeviceDescriptor {
-                        label: Some("Device descriptor"),
-                        features: wgpu::Features::ADDRESS_MODE_CLAMP_TO_BORDER,
-                        limits: wgpu::Limits::default(),
-                    },
-                    None,
-                ))
-                .expect("Failed to create device");
-
-                config.width = size.width;
-                config.height = size.height;
-                surface.configure(&device, &config);
-
-                let context = WgpuContext::new(&device, &queue, config.format, None, 1);
-
-                state = Some((window, surface, context, device, queue));
-            }
-
-            Event::Suspended => {
-                // Dump the state.
-                state = None;
-            }
-
-            Event::RedrawEventsCleared => {
-                if let Some((_, surface, context, device, queue)) = &mut state {
-                    let frame = surface
-                        .get_current_texture()
-                        .expect("Failed to get texture view");
-
-                    // Create the command encoder.
-                    let mut encoder = device
-                        .create_command_encoder(&wgpu::CommandEncoderDescriptor { label: None });
-
-                    let view = frame
-                        .texture
-                        .create_view(&wgpu::TextureViewDescriptor::default());
-
-                    {
-                        let mut render_pass =
-                            encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                                label: None,
-                                color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                                    view: &view,
-                                    resolve_target: None,
-                                    ops: wgpu::Operations {
-                                        load: wgpu::LoadOp::Clear(wgpu::Color {
-                                            r: 0.0,
-                                            g: 0.0,
-                                            b: 0.0,
-                                            a: 1.0,
-                                        }),
-                                        store: true,
-                                    },
-                                })],
-                                depth_stencil_attachment: None,
-                            });
-
-                        let mut piet = context.prepare(device, queue, window_size.0, window_size.1);
-                        draw(&mut piet, window_size.0, window_size.1);
-
-                        // Finish the render pass.
-                        context.render(&mut render_pass);
-                    }
-
-                    // Submit the command buffer to the queue.
-                    queue.submit(Some(encoder.finish()));
-
-                    // Present the frame.
-                    frame.present();
-                }
-            }
-
-            Event::WindowEvent { event, .. } => match event {
-                WindowEvent::CloseRequested => control_flow.set_exit(),
-                WindowEvent::Resized(size) => {
-                    window_size = (size.width, size.height);
-                    if let Some((_, surface, _, device, _)) = &state {
-                        config.width = size.width;
-                        config.height = size.height;
-                        surface.configure(device, &config);
-                    }
-                }
-                _ => {}
-            },
-
-            _ => {}
-        }
-    })
+    util::run(draw)
 }
 
 fn generate_five_pointed_star(center: Point, inner_radius: f64, outer_radius: f64) -> BezPath {
